requirements.txt
audioop-lts>=0.2.1  # Python 3.13+ replacement for built-in audioop

uvicorn[standard]>=0.24.0
websockets>=12.0
twilio>=8.10.0
deepgram-sdk==3.10.0
openai>=1.12.0
python-dotenv>=1.0.0
httpx>=0.26.0
pyngrok>=7.0.0